refactor: 129 - move fetching from layout
