+++ src/components/catalog/layout.ts
@@ -1,18 +1,10 @@
 import { createDiv } from '../../utils/create-elements/create-tags';
-import { getAllProducts } from '../../app/api';
-import { createProductList } from './product-list';
+import { fetchProductCards } from '../../helpers/fetch-product-cards';
 
 export async function catalogLayout(): Promise<HTMLDivElement> {
   const layout = createDiv({});
-  const data = await getAllProducts();
 
-  try {
-    if ('results' in data) {
-      const cards = createProductList(data);
-      layout.append(cards);
-    }
-  } catch {
-    throw new Error('Failed to fetch products');
-  }
+  await fetchProductCards(layout);
+
   return layout;
 }
